requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-ollama
sentence-transformers